Рефакторинг модуля с утилитами
