+++ src/model/points-model.js
@@ -1,5 +1,5 @@
 import {generatePoint} from '../mock/points.js';
-import {generateData} from '../utils.js';
+import {generateData} from '../utils/common.js';
 import { POINTS_AMOUNT } from '../const.js';
 
 export default class PointsModel {
